Initial commit
So this is where it all starts: The first commit of this project. It's not perfect by any stretch of the imagination, but it is "good enough" for an initial commit of this *learning* experience: "Rome was not built in a day either".
A quick overview of this project's structure:
"app.py" is the main container & controller of this Markdown editor, accessing "layout.py" for - well - building the GUI as well as the "libs" directory for the actual backend parsing and rendering logic.
"libs" includes the modules of the parser ("parser.py", "blocks.py", "inlines.py", "node.py", "regexps.py" and "link_reference_definition.py") and of the currently implemented renderers, "htmlrenderer.py" and "jsonrenderer.py".
"stylesheets" include the stylesheets used for styling the produced HTML in the rendering pane as well as in the exported HTML files (currently only "github-markdown.css" and "foghorn.css"), along with a "CREDITS.txt" providing the links to the locations these .css files were taken from - as well as providing the necessary licensing information.
And last - but not least - "windows.spec" and "linux.spec" in the root directory are the "build files" used to build this application into a standalone format using PyInstaller.