Hotfix crashes caused by static file path representations
Originally, the file paths to the stylesheets were constructed by calling os.getcwd() and concatenating a string in the form of /stylesheets/name.css to it. This would lead to the program crashing in Windows, as file paths are represented differently in Windows (\\ instead of /). Using os.path.join() resolves this issue.
Also, windows.spec has been edited to now produce a --onedir build, instead of bundling everything into one .exe file. This is a band-aid solution to the issue of the program trying to look for the stylesheets in a wrong location and consequently crashing.