Add parsing of inline code spans
For whatever reason, I completely forgot to include inline code spans into the inline parser. This has now been rectified, with all related tests of the CommonMark spec now passing (the ones not passing from this section - Examples no. 343-346 - are not failing because of the implementation of the inline parsing itself, but rather issues in other parts of the parser).
Also renamed the NodeType name from just CODE to INLINE_CODE for the sake of clarity (node.py).
Finally, edited the README.md to correct the number of CommonMark-tests now passing.