Fix HTML block & inline parsing
The parser now parses every HTML block & inline example of the CommonMark spec correctly (the tests still failing from these sections are not because of the HTML parsing implementation itself but rather other shortcomings of the parser). This was achieved - first and foremost - by modifing/correcting the regular expressions in "regexps.py". However, some small adjustments in parser.py were necessary as well.
Also removed some unneccessary things relating to HTML inlines, e.g. the "HTMLInlineNode" class and tag name capturing. This also necessitated some adjustments in the renderers.
Finally - of course - modified the README.md to reflect the change in the number of passing tests.