Refactor regexps.py
Because the regular expressions in "regexps.py" were - at times - unreadable/not understandable - and some were outright superfluous -, I took the time to refactor (almost) all regular expressions in this file. This - to my surprise - resulted in not only more readable code, but the parser passing 17 more tests than before (I'll take it though).
This revealed a shortcoming of the inline parser, which was consequently fixed. Also, "blocks.py" was modified to remove redundant/unneccessary elements from the last commit.
Finally - of course - modified the README.md to reflect the change in the number of passing tests.