Fix parsing of setext headings
This required some more in depth changes to the parser logic itself (e.g. a paragraph remaining open at the sight of a setext heading underline, so we can determine whether it is actually a setext heading, a thematic break, or just plain dashes). However, now (almost) all CommonMark spec tests regarding setext headings now pass (of course with the exception that there are failing tests in this section still due to shortcomings of the parser in other areas).
Of course updated README.md to reflect the change in passing tests.