Add syntax highlighting
Finally got around to add - albeit rudimentary - syntax highlighting for the editing pane. The biggest chunk of the time spent on this feature was not invested in the "algorithm" itself, but rather trying to debug some "RecursionError" I was getting, before finding out textEdit.blockSignals(True) solves all of my problems (three days down the drain >:( ).
There probably still is a metric ton of errors with this highlighting algorithm that I did not catch in my short testing period, however it is "good enough" for this initial commit of this feature.