Update UI & UX
Finally introduced a dark mode to the application (light mode can still be accessed; it is - however - not the default). The application stylesheets for the dark and light theme are not my own creation, but were taken from https://github.com/Alexhuszagh/BreezeStyleSheets (as credited in the stylesheet.qss). Also changed the status bar at the bottom to be prettier than before (the text was somewhat misaligned before these changes).
Also updated the UX such that if the user selected some text in the editing field and presses the shortcut for inserting italic or bold delimiters, it surrounds the selected text with said delimiters rather than deleting the text and then inserting the delimiters.